pyproject.toml
[project]
name = "genimage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers>=0.35.2",
    "huggingface-hub>=0.36.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "safetensors>=0.7.0",
    "scipy>=1.16.3",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.57.1",
]
